First functional draft of shiny app
Not sure why it takes so long to draw the image.
Just the R script takes about the same time, so I should look for issues there.
Need to change the inputs to be reactive so that only the permissible values are shown
(limited by inputs above, and making so assumptions about inputs below).